:bookmark: **Version 1.23.01**
:ok_hand: - Apresentação somente dos casos de teste que não estão com a label 'Execution::Only Manual'.
:ok_hand: - Remoção dos comentários desnecessários.
:ok_hand: - Correção da ortografia das frases do log.
:ok_hand: - Limpeza do arquivo de dicionário.
:ok_hand: - Aceitação do NO ou NÃO em maiúsculo.
:ok_hand: - Sublinhando a palavra "ignorado" no log (Novo template do Notepad++).
:sparkles: - Listagem dos status dos testes na tabela de listagem.
:sparkles: - Apresentação de cores para o status de cada caso de teste.
:bug: - Correção do erro, na execução, sem salvamento da evidência.
:bug: - Correção do erro 'remove: path should be string, bytes or os.PathLike, not NoneType' quando o teste era executado pela primeira vez.
:bug: - Correção do erro de criação do diretório de evidência da primeira execução.
:bug: - Correção da geração do print para o verbo alterar.
:bug: - Correção da validação do text dentro do Alerta.
:bug: - Correção da interrupção da automação ao não encontrar um componente esperado.

+++ Automation/modules/automationAux.py
@@ -8,7 +8,7 @@ logs = None
 class Textcolor:
     BLUE = '\033[1;34;47m'  # Blue
     HIGHLIGHT = '\033[7m'   # Highlight
-    GREEN = '\033[32m'      # Green (seems yellow)
+    GREEN = '\033[32m'      # Green
     WARNING = '\033[93m'    # Yellow
     FAIL = '\033[91m'       # Red
     END = '\033[00m'        # End of line
@@ -139,8 +139,6 @@ class Main:
 
             if not Lib.os.path.exists(path):
                 Lib.os.makedirs(path)
-                print(f"{Textcolor.FAIL}{logs['ErrorCreateDirectory']['Msg']}{Textcolor.END}")
-                # Main.addLogs(message="General", value=logs["ErrorCreateDirectory"], value1=str(path))
                 return True
             else:
                 # Clear the old logs and evidences (Older than 30 days).
@@ -236,8 +234,8 @@ class Main:
 
                 verb = step.split()[0]
                 # Don't execute the step with No / Não.
-                if verb not in ('"No"', '"Não"', '"No"'.replace('"', ''),
-                                '"Não"'.replace('"', '')):
+                if verb.upper() not in ('"NO"', '"NÃO"', '"NO"'.replace('"', ''),
+                                        '"NÃO"'.replace('"', '')):
 
                     # Last step or take_picture_status is true.
                     if verb not in ('Fechar', 'Cerrar', 'Close') and otherConfigs['API_Step'] is False:  # and take_picture_status:
@@ -484,7 +482,7 @@ class Main:
                         # Lib.os.remove(Lib.os.path.join(directories['TestSetPath'], item))
                         Lib.os.remove(Lib.os.path.join(folder_path, item))
 
-            else:
+            elif exact_file is not None:
                 Lib.os.remove(exact_file)
 
         except Exception as ex:
